Mock cleanup with deprecations
