+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomii/BlueMoveFoundation.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.autonomii;
 import android.graphics.Color;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -16,7 +17,11 @@ import org.firstinspires.ftc.teamcode.robotplus.hardware.MecanumDrive;
 import org.firstinspires.ftc.teamcode.robotplus.hardware.MotorPair;
 import org.firstinspires.ftc.teamcode.robotplus.hardware.Robot;
 
+/**
+ * @deprecated because we can't start in the triangle
+ */
 @Autonomous(name = "BlueMoveFoundation", group = "Blue")
+@Disabled
 public class BlueMoveFoundation extends LinearOpMode implements AutonomousConstants {
 
     private Robot robot;
