--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomii/EncoderTesting.java
@@ -1,6 +1,0 @@
-package org.firstinspires.ftc.teamcode.autonomii;
-
-public class EncoderTesting {
-
-
-}
