+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomii/RedMoveFoundation.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.autonomii;
 import android.graphics.Color;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -15,7 +16,12 @@ import org.firstinspires.ftc.teamcode.robotplus.autonomous.TimeOffsetVoltage;
 import org.firstinspires.ftc.teamcode.robotplus.hardware.MecanumDrive;
 import org.firstinspires.ftc.teamcode.robotplus.hardware.Robot;
 
+/**
+ * @deprecated
+ * @see RedMoveFoundationTriangle
+ */
 @Autonomous(name = "RedMoveFoundation", group = "Generic")
+@Disabled
 public class RedMoveFoundation extends LinearOpMode implements AutonomousConstants {
 
     private Robot robot;
