+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/DoubleCheck.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomii;
+package org.firstinspires.ftc.teamcode.lib;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.ColorSensor;
