+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/Test.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomii;
+package org.firstinspires.ftc.teamcode.test;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
